+++ backend-treasurehunt/src/main/java/com/geoschnitzel/treasurehunt/backend/api/HelloWorldApi.java
@@ -1,10 +1,10 @@
-package com.geoschnitzel.treasurehunt.backend;
+package com.geoschnitzel.treasurehunt.backend.api;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class HelloWorldController {
+public class HelloWorldApi {
 
     @GetMapping("/helloWorld")
     public String helloWorld() {
