+++ backend-treasurehunt/src/main/java/com/geoschnitzel/treasurehunt/backend/model/HintImage.java
@@ -3,7 +3,9 @@ package com.geoschnitzel.treasurehunt.backend.model;
 import javax.persistence.Entity;
 
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
+@EqualsAndHashCode(callSuper = true)
 @Entity
 @Data
 public class HintImage extends Hint {
@@ -11,5 +13,6 @@ public class HintImage extends Hint {
     private String imageFileName;
 
     private String mimeType;
+
 
 }
