+++ backend-treasurehunt/src/main/java/com/geoschnitzel/treasurehunt/backend/model/HintText.java
@@ -3,10 +3,13 @@ package com.geoschnitzel.treasurehunt.backend.model;
 import javax.persistence.Entity;
 
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
+@EqualsAndHashCode(callSuper = true)
 @Entity
 @Data
 public class HintText extends Hint {
 
     private String description;
+
 }
