+++ backend-treasurehunt/src/main/java/com/geoschnitzel/treasurehunt/backend/service/TestDataService.java
@@ -1,11 +1,14 @@
-package com.geoschnitzel.treasurehunt.backend;
+package com.geoschnitzel.treasurehunt.backend.service;
+
+
+import com.geoschnitzel.treasurehunt.backend.api.TestDataApi;
 
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RestController;
 
 @Service
 @RestController
-public class TestDataGenerator implements TestDataController {
+public class TestDataService implements TestDataApi {
 
     @Override
     public void generateTestData() {
