+++ backend-treasurehunt/src/main/resources/application.properties
@@ -1,2 +1,7 @@
 logging.level.org.springframework.web=INFO
-spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.properties.hibernate.use_sql_comments=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.type=trace
+spring.jpa.generate-ddl=true
+spring.jpa.hibernate.ddl-auto=create
